refactor(codec): extract analyzer into CodecAnalyzer.kt (step 3)
Move the KSP-declaration → IR analysis layer out of CodecEmitter: the 32
analyze/detect/resolve functions plus their pure helper closure (classNameOf,
readMessageWireOrder, parseWhenExpression, classifyVariantWireSize, …), the
WhenExpression parse type, and the peekable-kind sets — 55 declarations, ~2.9k
lines. The whole closure is free of emit/log state (no codeGenerator, no logger,
no build*/append* calls — verified), so it moves to package-internal top-level
functions and CodecEmitter (and the dispatch emitters) call them unqualified and
unchanged. No class/context plumbing needed.

CodecEmitter: 9408 → ~6.3k lines. Codegen byte-identical (CodecSnapshotTest green).
Step 3 of the incremental split.

No behavior change. Verified: :buffer-codec-processor:test :buffer-codec-test:jvmTest ktlintCheck.
